+++ CHANGELOG.md
@@ -1,6 +1,29 @@
+## v0.2.0 (2021-06-07)
+
+#### Breaking Changes
+
+- Replaced `icon/3` with `icon/2`, the `type` argument should be passed as an option instead
+
+#### Enhancements
+
+- Added `:type` to config for setting the default icon type
+
+## v0.1.4 (2021-05-29)
+
+- All options passed to `icon/3` will be added to the SVG tag as HTML attributes
+- Added `opts` prop to the Surface Icon component, which will be added to the SVG tag as HTML attributes
+
+## v0.1.3 (2021-05-24)
+
+- Improved documentation
+
+## v0.1.2 (2021-05-22)
+
+- Fixed Surface Icon component error when not passing the class prop
+
 ## v0.1.1 (2021-05-19)
 
-- Add Surface Icon component
+- Added Surface Icon component
 
 ## v0.1.0 (2021-05-19)